setting: Refactor sample code
