+++ myongji-graduate/app/utils/chart.util.ts
@@ -1,5 +1,4 @@
 // util layer sample code
-// helper.ts 말고도 style, test 등등 여러 util 함수들을 정의
 import { Revenue } from '../business/invoice/revenue.query';
 
 export const generateYAxis = (revenue: Revenue[]) => {
